Add HathiTrust crawler
